+++ boot/src/test/kotlin/integration/log/ProbeLogTest.kt
@@ -14,21 +14,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package integration
+package integration.log
 
+import integration.ProbeIntegrationTest
 import io.vertx.junit5.VertxTestContext
 import io.vertx.kotlin.coroutines.await
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Test
-import spp.protocol.instrument.LiveBreakpoint
-import spp.protocol.instrument.event.LiveBreakpointHit
+import spp.protocol.instrument.LiveLog
 import spp.protocol.instrument.event.LiveInstrumentEvent
 import spp.protocol.instrument.event.LiveInstrumentEventType
+import spp.protocol.instrument.event.LiveLogHit
 import spp.protocol.instrument.location.LiveSourceLocation
 
-class ProbeBreakpointTest : ProbeIntegrationTest() {
+class ProbeLogTest : ProbeIntegrationTest() {
 
     private fun doTest() {
         val a = 1
@@ -44,15 +45,14 @@ class ProbeBreakpointTest : ProbeIntegrationTest() {
 
     @Test
     fun testPrimitives(): Unit = runBlocking {
-        val breakpointId = "probe-breakpoint-test"
         val testContext = VertxTestContext()
-        getLiveInstrumentSubscription(breakpointId).handler {
+        val logId = "probe-log-test"
+        getLiveInstrumentSubscription(logId).handler {
             testContext.verify {
                 val event = LiveInstrumentEvent.fromJson(it.body())
-                if (event.eventType == LiveInstrumentEventType.BREAKPOINT_HIT) {
-                    val item = event as LiveBreakpointHit
-                    val vars = item.stackTrace.first().variables
-                    assertEquals(10, vars.size)
+                if (event.eventType == LiveInstrumentEventType.LOG_HIT) {
+                    val item = event as LiveLogHit
+                    assertEquals("1 a a", item.logResult.logs.first().toFormattedMessage())
 
                     testContext.completeNow()
                 }
@@ -61,19 +61,21 @@ class ProbeBreakpointTest : ProbeIntegrationTest() {
 
         assertNotNull(
             instrumentService.addLiveInstrument(
-                LiveBreakpoint(
+                LiveLog(
+                    logFormat = "{} {} {}",
+                    logArguments = listOf("a", "b", "c"),
                     location = LiveSourceLocation(
-                        source = ProbeBreakpointTest::class.java.name,
-                        line = 43,
+                        source = ProbeLogTest::class.java.name,
+                        line = 44,
                         service = "spp-test-probe"
                     ),
                     applyImmediately = true,
-                    id = breakpointId
+                    id = logId
                 )
             ).await()
         )
 
-        log.info("Triggering breakpoint")
+        log.info("Triggering log")
         doTest()
 
         errorOnTimeout(testContext)
